requirements.txt
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.6
Flask-RESTful==0.3.6
Jinja2==2.10
MarkupSafe==1.0
PyYAML==3.13
Werkzeug==0.14.1
aniso8601==3.0.2
itsdangerous==0.24
pytz==2018.5
six==1.11.0
